requirements.txt
paho-mqtt
meshtastic
cryptography
